# src/main/java/com/eaglebank/dto/ErrorResponse.java
package com.eaglebank.dto;

import lombok.Data;

@Data
public class ErrorResponse {
    private String message;
    
    public ErrorResponse(String message) {
        this.message = message;
    }
}